Extract ArcTests into separate test file; verify the 'spacing' visual generation as well
